[SSD-70] refactor: 패키지 구조 변경
- JWT 관련 패키지 auth 패키지 하위로 이동
