+++ backend/src/main/java/com/timepaper/backend/global/auth/jwt/filter/JwtAuthenticationFilter.java
@@ -1,6 +1,6 @@
-package com.timepaper.backend.global.jwt.filter;
+package com.timepaper.backend.global.auth.jwt.filter;
 
-import com.timepaper.backend.global.jwt.util.JWTUtil;
+import com.timepaper.backend.global.auth.jwt.util.JWTUtil;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
